+++ packages/gatsby-plugin-netlify-cms/src/cms-identity.js
@@ -2,11 +2,20 @@
 import netlifyIdentityWidget from "netlify-identity-widget"
 
 window.netlifyIdentity = netlifyIdentityWidget
+
+const addLoginListener = () =>
+  netlifyIdentityWidget.on(`login`, () => {
+    document.location.href = `${__PATH_PREFIX__}/${CMS_PUBLIC_PATH}/`
+  })
+
 netlifyIdentityWidget.on(`init`, user => {
   if (!user) {
-    netlifyIdentityWidget.on(`login`, () => {
-      document.location.href = `${__PATH_PREFIX__}/${CMS_PUBLIC_PATH}/`
+    addLoginListener()
+  } else {
+    netlifyIdentityWidget.on(`logout`, () => {
+      addLoginListener()
     })
   }
 })
+
 netlifyIdentityWidget.init()