+++ packages/gatsby-transformer-csv/src/__tests__/gatsby-node.js
@@ -30,12 +30,14 @@ describe(`Process  nodes correctly`, () => {
     const actions = { createNode, createParentChildLink }
     const createNodeId = jest.fn()
     createNodeId.mockReturnValue(`uuid-from-gatsby`)
+    const createContentDigest = jest.fn().mockReturnValue(`contentDigest`)
 
     await onCreateNode({
       node,
       loadNodeContent,
       actions,
       createNodeId,
+      createContentDigest,
     }).then(() => {
       expect(createNode.mock.calls).toMatchSnapshot()
       expect(createParentChildLink.mock.calls).toMatchSnapshot()
@@ -52,6 +54,7 @@ describe(`Process  nodes correctly`, () => {
     const actions = { createNode, createParentChildLink }
     const createNodeId = jest.fn()
     createNodeId.mockReturnValue(`uuid-from-gatsby`)
+    const createContentDigest = jest.fn().mockReturnValue(`contentDigest`)
 
     await onCreateNode(
       {
@@ -59,6 +62,7 @@ describe(`Process  nodes correctly`, () => {
         loadNodeContent,
         actions,
         createNodeId,
+        createContentDigest,
       },
       { noheader: true }
     ).then(() => {