+++ packages/gatsby/src/utils/api-node-docs.js
@@ -78,13 +78,41 @@ exports.createPages = true
 exports.createPagesStatefully = true
 
 /**
- * Extension point to tell plugins to source nodes.
+ * Extension point to tell plugins to source nodes. This API is called during
+ * the Gatsby bootstrap sequence. Source plugins use this hook to create nodes.
+ * This API is called exactly once per plugin (and once for your site's
+ * `gatsby-config.js` file). If you define this hook in `gatsby-node.js` it
+ * will be called exactly once after all of your source plugins have finished
+ * creating nodes.
  *
  * See also the documentation for [`createNode`](/docs/actions/#createNode).
  * @example
- * exports.sourceNodes = ({ actions }) => {
+ * exports.sourceNodes = ({ actions, createNodeId, createContentDigest }) => {
  *   const { createNode } = actions
- *   // Create nodes here.
+ *
+ *   // Data can come from anywhere, but for now create it manually
+ *   const myData = {
+ *     key: 123,
+ *     foo: `The foo field of my node`,
+ *     bar: `Baz`
+ *   }
+ *
+ *   const nodeContent = JSON.stringify(myData)
+ *
+ *   const nodeMeta = {
+ *     id: createNodeId(`my-data-${myData.key}`),
+ *     parent: null,
+ *     children: [],
+ *     internal: {
+ *       type: `MyNodeType`,
+ *       mediaType: `text/html`,
+ *       content: nodeContent,
+ *       contentDigest: createContentDigest(myData)
+ *     }
+ *   }
+ *
+ *   const node = Object.assign({}, myData, nodeMeta)
+ *   createNode(node)
  * }
  */
 exports.sourceNodes = true
@@ -150,8 +178,8 @@ exports.onCreatePage = true
  *         args: {
  *           myArgument: {
  *             type: GraphQLString,
- *          }
- *         }
+ *           }
+ *         },
  *         resolve: (source, fieldArgs) => {
  *           return `Id of this node is ${source.id}.
  *                   Field was called with argument: ${fieldArgs.myArgument}`
@@ -250,3 +278,17 @@ exports.onPostBuild = true
  * See gatsby-transformer-remark and gatsby-source-contentful for examples.
  */
 exports.onPreExtractQueries = true
+
+/**
+ * Run when gatsby develop server is started, its useful to add proxy and middleware
+ * to the dev server app
+ * @param {object} $0
+ * @param {Express} $0.app The [Express app](https://expressjs.com/en/4x/api.html#app) used to run the dev server
+ * @example
+ * exports.onCreateDevServer = ({ app }) => {
+ *   app.get('/hello', function (req, res) {
+ *     res.send('hello world')
+ *   })
+ * }
+ */
+exports.onCreateDevServer = true