+++ www/src/pages/plugins.js
@@ -60,12 +60,11 @@ class Plugins extends Component {
                 }}
               >
                 Please use the search bar to find plugins that will make your
-                blazing fast site even more awesome. If you{`'`}d like to create
-                your own plugin, see the
+                blazing fast site even more awesome. If you
+                {`'`}d like to create your own plugin, see the
                 {` `}
                 <Link to="/docs/plugin-authoring/">Plugin Authoring</Link> page
-                in the docs!
-                To learn more about Gatsby plugins, visit the
+                in the docs! To learn more about Gatsby plugins, visit the
                 {` `}
                 <Link to="/docs/plugins">plugins doc page</Link>.
               </p>